Split out launch parameters.
