+++ torch/csrc/jit/codegen/cuda/executor.cpp
@@ -18,34 +18,6 @@ namespace jit {
 namespace fuser {
 namespace cuda {
 
-void LaunchParams::bind(int64_t val, ParallelType p_dim) {
-  switch (p_dim) {
-    case (ParallelType::TIDx):
-      checkAndSet(val, bdimx_, "blockDim.x");
-      break;
-    case (ParallelType::BIDx):
-      checkAndSet(val, gdimx_, "gridDim.x");
-      break;
-    case (ParallelType::TIDy):
-      checkAndSet(val, bdimy_, "blockDim.y");
-      break;
-    case (ParallelType::BIDy):
-      checkAndSet(val, gdimy_, "gridDim.y");
-      break;
-    case (ParallelType::TIDz):
-      checkAndSet(val, bdimz_, "blockdim.z");
-      break;
-    case (ParallelType::BIDz):
-      checkAndSet(val, gdimz_, "gridDim.z");
-      break;
-    default:
-      TORCH_INTERNAL_ASSERT(
-          false,
-          "Tried to bind invalid parallel type in launch config: ",
-          p_dim);
-  }
-}
-
 // Push a tensor to the arguments
 void KernelArgumentHolder::push(const at::Tensor& tensor) {
   changed = true;
